Add an option to force pre-O apps to use full screen aspect ratio
When an app target pre-O releases, the default max aspect ratio
is 1.86:1 which leads to ugly black areas on devices that have
screens with higher aspect ratio (for example Galaxy S8/S9).

This change adds an option to allow users to change default
aspect ratio for pre-O apps to 2.1 which would fit recent devices.

Wight554:
Simplify the code, we don't need the Settings for that, just bool

Change-Id: I88035fb079bad9a66b36fe5e861017af568b8f4c
Signed-off-by: Kshitij Gupta <kshitijgm@gmail.com>